refactor(toolkits): merge 4 GELLO scripts into a single gello.py with subcommands
Per reviewer: collapse the four single-purpose GELLO helper scripts
into one entry point with argparse subcommands so users see all the
related tools in one place and shared setup code (port resolution,
controller / expert bringup, safe reset_joint with actionable errors)
lives in exactly one copy.

  python toolkits/realworld_check/gello.py align-check
  python toolkits/realworld_check/gello.py align-sequential
  python toolkits/realworld_check/gello.py calibrate
  python toolkits/realworld_check/gello.py reset-to-gello

Removes:
  toolkits/realworld_check/gello_align_check.py
  toolkits/realworld_check/gello_align_sequential.py
  toolkits/realworld_check/gello_calibrate.py
  toolkits/realworld_check/gello_reset_to_gello.py
(net -400 lines after de-duping the 4 copies of _resolve_local_robotiq_port,
 setup_hardware, colour, and safe_reset_to.)

Updates the three shell wrappers in examples/embodiment/ to invoke the
new subcommand entry, and updates dual_franka.rst (EN + ZH) references.
