+++ examples/embodiment/gello_align.sh
@@ -23,7 +23,7 @@ set -euo pipefail
 
 EMBODIED_PATH="$( cd "$(dirname "${BASH_SOURCE[0]}")" && pwd )"
 REPO_PATH="$(dirname "$(dirname "$EMBODIED_PATH")")"
-SRC_FILE="${REPO_PATH}/toolkits/realworld_check/gello_align_sequential.py"
+SRC_FILE="${REPO_PATH}/toolkits/realworld_check/gello.py"
 
 export PYTHONPATH="${REPO_PATH}:${PYTHONPATH:-}"
 export FRANKA_ROBOT_IP="${FRANKA_ROBOT_IP:-172.16.0.2}"
@@ -40,4 +40,4 @@ if [ ! -f "${SRC_FILE}" ]; then
     exit 1
 fi
 
-exec python "${SRC_FILE}"
+exec python "${SRC_FILE}" align-sequential
